other windows & add vpn icon in bar
