+++ ignis/modules/__init__.py
@@ -1,0 +1,8 @@
+from .bar import Bar
+from .control_center import ControlCenter
+from .launcher import Launcher
+from .notification_popup import NotificationPopup
+from .osd import OSD
+from .powermenu import Powermenu
+
+__all__ = ["Bar", "ControlCenter", "Launcher", "NotificationPopup", "OSD", "Powermenu"]
