+++ ignis/modules/control_center/widgets/__init__.py
@@ -1,0 +1,15 @@
+from .quick_settings import QuickSettings
+from .brightness import Brightness
+from .volume import VolumeSlider
+from .user import User
+from .media import Media
+from .notification_center import NotificationCenter
+
+__all__ = [
+    "QuickSettings",
+    "Brightness",
+    "VolumeSlider",
+    "User",
+    "Media",
+    "NotificationCenter",
+]
