+++ ignis/modules/control_center/widgets/media.scss
@@ -1,3 +1,4 @@
+// JINJA2 TEMPLATE
 // WARNING: Do not import this file in style.scss
 
 .media-{{ desktop_entry }} {
