+++ ignis/modules/control_center/widgets/quick_settings/__init__.py
@@ -1,0 +1,3 @@
+from .quick_settings import QuickSettings
+
+__all__ = ["QuickSettings"]
