+++ ignis/modules/shared_widgets/__init__.py
@@ -1,3 +1,4 @@
 from .toggle_box import ToggleBox
+from .notification import NotificationWidget
 
-__all__ = ["ToggleBox"]
+__all__ = ["ToggleBox", "NotificationWidget"]
