+++ ignis/modules/shared_widgets/notification.py
@@ -1,9 +1,6 @@
 from ignis.widgets import Widget
-from ignis.services.notifications import Notification, NotificationService
+from ignis.services.notifications import Notification
 from ignis.utils import Utils
-from gi.repository import GLib  # type: ignore
-
-notifications = NotificationService.get_default()
 
 
 class ScreenshotLayout(Widget.Box):
@@ -138,103 +135,3 @@ class NotificationWidget(Widget.Box):
             css_classes=["notification"],
             child=[layout],
         )
-
-
-class Popup(Widget.Revealer):
-    def __init__(self, notification: Notification, **kwargs):
-        widget = NotificationWidget(notification)
-        super().__init__(child=widget, transition_type="slide_down", **kwargs)
-
-        notification.connect("closed", lambda x: self.destroy())
-
-    def destroy(self):
-        self.reveal_child = False
-        Utils.Timeout(self.transition_duration, self.unparent)
-
-
-def loading_notif_label() -> Widget.Label:
-    return Widget.Label(
-        label="Loading notifications...",
-        valign="center",
-        vexpand=True,
-        css_classes=["notification-center-info-label"],
-    )
-
-
-def no_notifications_label() -> Widget.Label:
-    return Widget.Label(
-        label="No notifications",
-        valign="center",
-        vexpand=True,
-        visible=notifications.bind("notifications", lambda value: len(value) == 0),
-        css_classes=["notification-center-info-label"],
-    )
-
-
-def on_notified(box: Widget.Box, notification: Notification) -> None:
-    notify = Popup(notification)
-    box.prepend(notify)
-    notify.reveal_child = True
-
-
-def load_notifications() -> list[Widget.Label | Popup]:
-    widgets = []
-    for i in notifications.notifications:
-        GLib.idle_add(lambda i=i: widgets.append(Popup(i, reveal_child=True)))
-    return widgets
-
-
-def notification_list() -> Widget.Box:
-    box = Widget.Box(
-        vertical=True,
-        child=[loading_notif_label()],
-        vexpand=True,
-        css_classes=["rec-unset"],
-        setup=lambda self: notifications.connect(
-            "notified",
-            lambda x, notification: on_notified(self, notification),
-        ),
-    )
-
-    Utils.ThreadTask(
-        load_notifications,
-        lambda result: box.set_child(result + [no_notifications_label()]),
-    ).run()
-
-    return box
-
-
-def notification_center() -> Widget.Box:
-    return Widget.Box(
-        vertical=True,
-        vexpand=True,
-        css_classes=["notification-center"],
-        child=[
-            Widget.Box(
-                css_classes=["notification-center-header", "rec-unset"],
-                child=[
-                    Widget.Label(
-                        label=notifications.bind(
-                            "notifications", lambda value: str(len(value))
-                        ),
-                        css_classes=["notification-count"],
-                    ),
-                    Widget.Label(
-                        label="notifications",
-                        css_classes=["notification-header-label"],
-                    ),
-                    Widget.Button(
-                        child=Widget.Label(label="Clear all"),
-                        halign="end",
-                        hexpand=True,
-                        on_click=lambda x: notifications.clear_all(),
-                        css_classes=["notification-clear-all"],
-                    ),
-                ],
-            ),
-            Widget.Scroll(
-                child=notification_list(),
-                vexpand=True,
-            ),
-        ],
-    )
