+++ ignis/scss/control_center.scss
@@ -154,4 +154,16 @@
         margin-left: 0.5rem;
         margin-right: 1rem;
     }
+
+    .volume-mainbox-speaker {
+        margin-top: 1rem;
+    }
+
+    .volume-mainbox-microphone {
+        margin-top: 0.25rem;
+    }
+
+    .volume-menubox-speaker {
+        margin-bottom: 0.75rem;
+    }
 }
